+++ src/Day17/Services/Instructions/Cdv.cs
@@ -20,6 +20,6 @@ public static  class Cdv
         var numerator = registerAValue;
         var denominator = Math.Pow(2, (uint)comboOperand);
 
-        return (uint)(int)(numerator / denominator);
+        return (uint)(numerator / denominator);
     }
 }
